refactor: rename mock api
